requirements.txt
Flask==2.0.1
Flask-WTF==0.15.1
pandas==1.3.3
scikit-learn==0.24.2
python-dotenv==0.19.0